sys-kernel/git-sources: update

+++ sys-kernel/git-sources/git-sources-3.0.34.ebuild
@@ -1,8 +1,9 @@
 # Copyright 1999-2012 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: sys-kernel/git-sources/git-sources-3.0.30.ebuild v1.1 2011/11/23 -tclover Exp $
+# $Header: bar-overlay/sys-kernel/git-sources/git-sources-3.0.34.ebuild,v 1.2 2012/06/13 22:22:25 -tclover Exp $
 
-EAPI=2
+EAPI=4
+
 UNIPATCH_STRICTORDER="yes"
 K_NOUSENAME="yes"
 K_NOSETEXTRAVERSION="yes"
@@ -24,26 +25,27 @@ DESCRIPTION="The very latest linux-stable.git, -git as pulled by git of the stab
 HOMEPAGE="http://www.kernel.org"
 EGIT_REPO_URI="git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git"
 EGIT_PROJECT=${PN}
-EGIT_TAG=v${PV/-r[0-9]*}
+EGIT_TAG=v${PV/%.0}
 EGIT_NOUNPACK="yes"
 
 EGIT_REPO_AUFS="git://aufs.git.sourceforge.net/gitroot/aufs/aufs${KV_MAJOR}-standalone.git"
 KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~ppc ~ppc64 ~sh ~sparc ~x86"
-IUSE="aufs bfs fbcondecor ck hz tuxonice"
+IUSE="aufs bfs fbcondecor ck hz"
+REQUIRED_USE="ck? ( bfs hz ) hz? ( || ( bfs ck ) )"
 
-BFS_FILE=${KV_MAJOR}.${KV_MINOR}-sched-bfs-406.patch
-BFS_URI=http://ck.kolivas.org/patches/bfs/${KV_MAJOR}.${KV_MINOR}.0/
-CK_VERSION=${KV_MAJOR}.${KV_MINOR}.0-ck1
-CK_FILE=${CK_VERSION}-broken-out.tar.bz2
-CK_URI="http://ck.kolivas.org/patches/${KV_MAJOR}.${KV_MINOR}/${KV_MAJOR}.${KV_MINOR}/${CK_VERSION}/"
-GEN_FILE=genpatches-${KV_MAJOR}.${KV_MINOR}-${K_GENPATCHES_VER}.extras.tar.bz2
-TOI_FILE=current-tuxonice-for-${KV_MAJOR}.${KV_MINOR}.patch.bz2
-SRC_URI="tuxonice?  ( http://tuxonice.net/files/${TOI_FILE} )
-		fbcondecor? ( http://dev.gentoo.org/~mpagano/genpatches/tarballs/${GEN_FILE} )
-		bfs?        ( ${BFS_URI}/${BFS_FILE/sched-bfs-406/bfs-406-413} ${CK_URI}/${CK_FILE} )
-		ck?         ( ${BFS_URI}/${BFS_FILE/sched-bfs-406/ck1-bfs-406-413} ${CK_URI}/${CK_FILE} )
-		hz?         ( ${CK_URI}/${CK_FILE} )
+bv=406
+bfs_src=${KV_MAJOR}.${KV_MINOR}-sched-bfs-${bv}.patch
+bfs_uri=http://ck.kolivas.org/patches/bfs/${KV_MAJOR}.${KV_MINOR}/
+cv=${KV_MAJOR}.${KV_MINOR}.0-ck1
+ck_src=${cv}-broken-out.tar.bz2
+ck_uri="http://ck.kolivas.org/patches/${KV_MAJOR}.0/${KV_MAJOR}.${KV_MINOR}/${cv}/"
+gen_src=genpatches-${KV_MAJOR}.${KV_MINOR}-${K_GENPATCHES_VER}.extras.tar.bz2
+SRC_URI="fbcondecor? ( http://dev.gentoo.org/~mpagano/genpatches/tarballs/${gen_src} )
+		bfs? ( ${ck_uri}/${ck_src} )
+		ck? ( ${ck_uri}/${ck_src} )
+		hz? ( ${ck_uri}/${ck_src} )
 "
+unset bfs_uri bv ck_uri cv
 
 K_EXTRAEINFO="This kernel is not supported by Gentoo due to its (unstable and)
 experimental nature. If you have any issues, try disabling a few USE flags
@@ -52,7 +54,7 @@ based on the latest mainline (stable) tree."
 
 src_unpack() {
 	git-2_src_unpack
-	use aufs && {
+	if use aufs; then
 		EGIT_BRANCH=aufs${KV_MAJOR}.${KV_MINOR}
 		unset EGIT_COMMIT
 		unset EGIT_TAG
@@ -61,40 +63,41 @@ src_unpack() {
 		export EGIT_SOURCEDIR="${WORKDIR}"/aufs${KV_MAJOR}-standalone
 		export EGIT_PROJECT=aufs${KV_MAJOR}-standalone
 		git-2_src_unpack
-	}
-	{ use bfs || use hz || use ck; } && unpack ${CK_FILE} || die "eek!"
+	fi
+	if use bfs || use hz || use ck; then
+		unpack ${ck_src} || die
+	fi
 }
 
 src_prepare() {
-	use aufs && {
-		for fle in Documentation fs include/linux/aufs_type.h; do
-			cp -pPR "${WORKDIR}"/aufs${KV_MAJOR}-standalone/$fle . || die "eek!"
+	if use aufs; then
+		pushd ../aufs${KV_MAJOR}-standalone && epatch \
+			"${FILESDIR}"/aufs-${KV_MAJOR}.${KV_MINOR}.6-loopback_fix.patch && popd
+		for file in Documentation fs include/linux/aufs_type.h; do
+			cp -pPR "${WORKDIR}"/aufs${KV_MAJOR}-standalone/$file . || die
 		done
-		mv aufs_type.h include/linux/ || die "eek!"
-		kernel_is ge 3 0 6 && cd ../aufs* && epatch "${FILESDIR}"/aufs-3.0.6-8-loopback.patch && cd ../linux-*
-		local AUFS_PREFIX=aufs${KV_MAJOR}-standalone/aufs${KV_MAJOR}
-		epatch "${WORKDIR}"/${AUFS_PREFIX}-{kbuild,base,standalone,loopback,proc_map}.patch
-	}
-	use fbcondecor && epatch "${DISTDIR}"/${GEN_FILE}
-	use tuxonice && epatch "${DISTDIR}"/${TOI_FILE}
+		mv aufs_type.h include/linux/ || die
+		local ap=aufs${KV_MAJOR}-standalone/aufs${KV_MAJOR}
+		epatch "${WORKDIR}"/${ap}-{kbuild,base,standalone,loopback,proc_map}.patch
+		epatch "${FILESDIR}"/aufs-${KV_MAJOR}.${KV_MINOR}-fix_devcgroup.patch
+	fi
+	use fbcondecor && epatch "${DISTDIR}"/${gen_src}
+	if use bfs || use ck; then
+		pushd ../patches && epatch "${FILESDIR}"/bfs-406-413.patch && popd
+	fi
 	if use ck; then
-		kernel_is ge 3 0 6 && cd ../patches && epatch "${FILESDIR}"/bfs-406-413.patch && cd ../linux-*
-		sed -i -e "s:ck1-version.patch::g" ../patches/series || die "eek!"
-		for pch in $(< ../patches/series); do epatch ../patches/$pch || die "eek!"; done
-		epatch "${DISTDIR}"/${BFS_FILE/sched-bfs-406/ck1-bfs-406-413}
-	else
-		use bfs && {
-			kernel_is gt 3 0 6 && cd ../patches && epatch "${FILESDIR}"/bfs-406-413.patch && cd ../linux-*
-			epatch ../patches/{${BFS_FILE},cpufreq-bfs_tweaks.patch}
-			epatch "${DISTDIR}"/${BFS_FILE/sched-bfs-406/bfs-406-413}
-		}
-		use hz && {
-			for pch in $(grep hz ../patches/series); do epatch ../patches/$pch || die "eek!"; done
-			epatch ../patches/preempt-desktop-tune.patch || die "eek!"
-		}
+		sed -i -e "s:ck1-version.patch::g" ../patches/series || die
+		for pch in $(< ../patches/series); do epatch ../patches/$pch || die; done
+	fi
+	use bfs && epatch ../patches/${bfs_src}
+	if use hz; then
+		for pch in $(grep hz ../patches/series); do 
+			epatch ../patches/$pch || die
+		done
+		epatch ../patches/preempt-desktop-tune.patch || die
 	fi
 	rm -r .git
-	sed -e "s:EXTRAVERSION =:EXTRAVERSION = -git:" -i Makefile || die "eek!"
+	sed -e "s:EXTRAVERSION =:EXTRAVERSION = -git:" -i Makefile || die
 }
 
 pkg_postinst() {
