+++ sys-kernel/git-sources/git-sources-3.4.2.ebuild
@@ -1,8 +1,9 @@
 # Copyright 1999-2012 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: sys-kernel/git-sources/git-sources-3.4.0.ebuild v1.1 2012/05/08 -tclover Exp $
+# $Header: bar-overlay/sys-kernel/git-sources/git-sources-3.4.2.ebuild,v 1.2 2012/06/13 22:22:25 -tclover Exp $
 
-EAPI=2
+EAPI=4
+
 UNIPATCH_STRICTORDER="yes"
 K_NOUSENAME="yes"
 K_NOSETEXTRAVERSION="yes"
@@ -30,20 +31,19 @@ EGIT_NOUNPACK="yes"
 EGIT_REPO_AUFS="git://aufs.git.sourceforge.net/gitroot/aufs/aufs${KV_MAJOR}-standalone.git"
 KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ia64 ~ppc ~ppc64 ~sh ~sparc ~x86"
 IUSE="aufs bfs fbcondecor ck hz"
+REQUIRED_USE="ck? ( bfs hz ) hz? ( || ( bfs ck ) )"
 
-BFS_VERSION=422
-BFS_FILE=${KV_MAJOR}.${KV_MINOR}-sched-bfs-${BFS_VERSION}.patch
-BFS_URI=http://ck.kolivas.org/patches/bfs/${KV_MAJOR}.${KV_MINOR}/
-CK_VERSION=${KV_MAJOR}.${KV_MINOR}-ck1
-CK_FILE=${CK_VERSION}-broken-out.tar.bz2
-CK_URI="http://ck.kolivas.org/patches/${KV_MAJOR}.0/${KV_MAJOR}.${KV_MINOR}/${CK_VERSION}/"
-GEN_FILE=genpatches-${KV_MAJOR}.${KV_MINOR}-${K_GENPATCHES_VER}.extras.tar.bz2
-TOI_FILE=current-tuxonice-for-${KV_MAJOR}.patch.bz2
-SRC_URI="fbcondecor? ( http://dev.gentoo.org/~mpagano/genpatches/tarballs/${GEN_FILE} )
-		bfs? 		( ${CK_URI}/${CK_FILE} )
-		ck?			( ${CK_URI}/${CK_FILE} )
-		hz? 		( ${CK_URI}/${CK_FILE} )
+bv=422
+bfs_src=${KV_MAJOR}.${KV_MINOR}-sched-bfs-${bv}.patch
+bfs_uri=http://ck.kolivas.org/patches/bfs/${KV_MAJOR}.${KV_MINOR}/
+cv=${KV_MAJOR}.${KV_MINOR}-ck1
+ck_src=${cv}-broken-out.tar.bz2
+ck_uri="http://ck.kolivas.org/patches/${KV_MAJOR}.0/${KV_MAJOR}.${KV_MINOR}/${cv}/"
+gen_src=genpatches-${KV_MAJOR}.${KV_MINOR}-${K_GENPATCHES_VER}.extras.tar.bz2
+SRC_URI="fbcondecor? ( http://dev.gentoo.org/~mpagano/genpatches/tarballs/${gen_src} )
+	bfs? ( ${ck_uri}/${ck_src} ) ck? ( ${ck_uri}/${ck_src} ) hz? ( ${ck_uri}/${ck_src} )
 "
+unset bfs_uri bv ck_uri cv
 
 K_EXTRAEINFO="This kernel is not supported by Gentoo due to its (unstable and)
 experimental nature. If you have any issues, try disabling a few USE flags
@@ -52,7 +52,7 @@ based on the latest mainline (stable) tree."
 
 src_unpack() {
 	git-2_src_unpack
-	use aufs && {
+	if use aufs; then
 		EGIT_BRANCH=aufs${KV_MAJOR}.${KV_MINOR}
 		unset EGIT_COMMIT
 		unset EGIT_TAG
@@ -61,33 +61,37 @@ src_unpack() {
 		export EGIT_SOURCEDIR="${WORKDIR}"/aufs${KV_MAJOR}-standalone
 		export EGIT_PROJECT=aufs${KV_MAJOR}-standalone
 		git-2_src_unpack
-	}
-	{ use bfs || use hz || use ck; } && unpack ${CK_FILE} || die "eek!"
+	fi
+	if use bfs || use hz || use ck; then
+		unpack ${ck_src} || die
+	fi
 }
 
 src_prepare() {
-	use aufs && {
-		for fle in Documentation fs include/linux/aufs_type.h; do
-			cp -pPR "${WORKDIR}"/aufs${KV_MAJOR}-standalone/$fle . || die "eek!"
+	if use aufs; then
+		for file in Documentation fs include/linux/aufs_type.h; do
+			cp -pPR "${WORKDIR}"/aufs${KV_MAJOR}-standalone/$file . || die
 		done
-		mv aufs_type.h include/linux/ || die "eek!"
-		local AUFS_PREFIX=aufs${KV_MAJOR}-standalone/aufs${KV_MAJOR}
-		epatch "${WORKDIR}"/${AUFS_PREFIX}-{kbuild,base,standalone,loopback,proc_map}.patch
-	}
-	use fbcondecor && epatch "${DISTDIR}"/${GEN_FILE}
+		mv aufs_type.h include/linux/ || die
+		local ap=aufs${KV_MAJOR}-standalone/aufs${KV_MAJOR}
+		epatch "${WORKDIR}"/${ap}-{kbuild,base,standalone,loopback,proc_map}.patch
+	fi
+	use fbcondecor && epatch "${DISTDIR}"/${gen_src}
 	if use ck; then
-		sed -i -e "s:ck1-version.patch::g" ../patches/series || die "eek!"
-		for pch in $(< ../patches/series); do epatch ../patches/$pch || die "eek!"; done
-	else
-		use bfs && epatch ../patches/${BFS_FILE}
-		use hz && {
-			for pch in $(grep hz ../patches/series)
-			do epatch ../patches/$pch || die "eek!"; done
-			epatch ../patches/preempt-desktop-tune.patch || die "eek!"
-		}
+		sed -i -e "s:ck1-version.patch::g" ../patches/series || die
+		for pch in $(< ../patches/series); do
+			epatch ../patches/$pch || die
+		done
+	fi
+	use bfs && epatch ../patches/${bfs_src}
+	if use hz; then
+		for pch in $(grep hz ../patches/series); do 
+			epatch ../patches/$pch || die
+		done
+		epatch ../patches/preempt-desktop-tune.patch || die
 	fi
 	rm -r .git
-	sed -e "s:EXTRAVERSION =:EXTRAVERSION = -git:" -i Makefile || die "eek!"
+	sed -e "s:EXTRAVERSION =:EXTRAVERSION = -git:" -i Makefile || die
 }
 
 pkg_postinst() {
